requirements-lock.txt
# core runtime
pydantic==2.8.2
structlog==24.4.0
PyYAML==6.0.2
cryptography==43.0.1
numpy==1.26.4
orjson==3.10.7
click==8.1.7

# api
fastapi==0.112.2
uvicorn==0.30.6
pydantic-settings==2.4.0

# embeddings & models (CPU default)
transformers==4.44.2
tokenizers==0.19.1
torch==2.3.1
sentencepiece==0.2.0

# storage (relational default; vector optional)
sqlalchemy==2.0.32
duckdb==1.1.0
lancedb==0.7.4
pyarrow==16.1.0

# testing & dev
pytest==8.3.2
pytest-cov==5.0.0
ruff==0.6.5